4.1.6 Отрефакторит модуль с утилитами
