+++ src/mocks/event-mock.js
@@ -1,4 +1,4 @@
-import { getRandomInteger, getRandomArrayItem } from '../utils/utils.js';
+import { getRandomInteger, getRandomArrayItem } from '../utils/common.js';
 import { generateDescription, generateEventType, generatePhotos, generateDateFrom, generateDateTo, generateOffers } from './utils-mock.js';
 import { DESTINATIONS } from './data-mock';
